PetLover/requirements.txt
boto==2.48.0
certifi==2018.1.18
cffi==1.11.5
colorama==0.3.9
colorlog==3.1.2
Django==2.0.3
django-bootstrap3==9.1.0
django-braces==1.12.0
django-debug-toolbar==1.9.1
django-markwhat==1.6.0
keyring==5.7.1
markdown2==2.3.5
MarkupSafe==1.0
mistune==0.8.3
pycparser==2.18
pytz==2018.3
six==1.11.0
sqlparse==0.2.4
tornado==5.0.1
wincertstore==0.2
